+++ add-custom-dynamic-entity-data-source/README.metadata.json
@@ -1,2 +1,46 @@
 {
+    "category": "Layers",
+    "description": "Create a custom dynamic entity data source and display it using a dynamic entity layer.",
+    "formal_name": "AddCustomDynamicEntityDataSource",
+    "ignore": false,
+    "images": [
+        "add-custom-dynamic-entity-data-source.png"
+    ],
+    "keywords": [
+        "MutableSharedFlow",
+        "SharedFlow",
+        "data",
+        "dynamic",
+        "entity",
+        "flow",
+        "label",
+        "labeling",
+        "live",
+        "real-time",
+        "stream",
+        "track",
+        "CustomDynamicEntityDataSource.EntityFeedProvider",
+        "DynamicEntity",
+        "DynamicEntityDataSource",
+        "DynamicEntityLayer",
+        "LabelDefinition",
+        "TrackDisplayProperties"
+    ],
+    "language": "kotlin",
+    "redirect_from": "",
+    "relevant_apis": [
+        "CustomDynamicEntityDataSource.EntityFeedProvider",
+        "DynamicEntity",
+        "DynamicEntityDataSource",
+        "DynamicEntityLayer",
+        "LabelDefinition",
+        "TrackDisplayProperties"
+    ],
+    "snippets": [
+        "/src/main/java/com/esri/arcgismaps/sample/addcustomdynamicentitydatasource/MainActivity.kt",
+        "/src/main/java/com/esri/arcgismaps/sample/addcustomdynamicentitydatasource/components/CustomEntityFeedProvider.kt",
+        "/src/main/java/com/esri/arcgismaps/sample/addcustomdynamicentitydatasource/components/MapViewModel.kt",
+        "/src/main/java/com/esri/arcgismaps/sample/addcustomdynamicentitydatasource/screens/MainScreen.kt"
+    ],
+    "title": "Add custom dynamic entity data source"
 }
